Break functions into more modules to separate the functionality, update the readme page.
